--- Source/ABLASTR/Make.package
@@ -1,1 +1,0 @@
-VPATH_LOCATIONS   += $(WARPX_HOME)/Source/ABLASTR
