+++ Source/ablastr/profiler/Make.package
@@ -1,0 +1,3 @@
+#CEXE_sources += ParticleBoundaries.cpp
+
+VPATH_LOCATIONS   += $(WARPX_HOME)/Source/ablastr/profiler
